# Web/gin-gorm-oj/docs/swagger.yaml
info:
  contact: {}
  description: 这里是描述
  title: 项目名称
  version: "1.0"
paths:
  /admin/category-create:
    post:
      parameters:
      - description: token
        in: header
        name: authorization
        required: true
        type: string
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: parentId
        in: formData
        name: parentId
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 分类创建
      tags:
      - 私有方法
  /admin/category-delete:
    delete:
      parameters:
      - description: token
        in: header
        name: authorization
        required: true
        type: string
      - description: identity
        in: query
        name: identity
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 分类删除
      tags:
      - 私有方法
  /admin/category-list:
    get:
      parameters:
      - description: token
        in: header
        name: authorization
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: keyword
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 分类列表
      tags:
      - 私有方法
  /admin/category-modify:
    put:
      parameters:
      - description: token
        in: header
        name: authorization
        required: true
        type: string
      - description: identity
        in: formData
        name: identity
        required: true
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: parentId
        in: formData
        name: parentId
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 分类修改
      tags:
      - 私有方法
  /admin/problem-create:
    post:
      parameters:
      - description: token
        in: header
        name: authorization
        required: true
        type: string
      - description: title
        in: formData
        name: title
        required: true
        type: string
      - description: content
        in: formData
        name: content
        required: true
        type: string
      - description: max_runtime
        in: formData
        name: max_runtime
        type: integer
      - description: max_memory
        in: formData
        name: max_memory
        type: integer
      - description: category_ids
        in: formData
        name: category_ids
        type: array
      - description: test_cases
        in: formData
        name: test_cases
        required: true
        type: array
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 问题创建
      tags:
      - 私有方法
  /login:
    post:
      parameters:
      - description: username
        in: formData
        name: username
        type: string
      - description: password
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 用户登陆
      tags:
      - 公共方法
  /problemDetail:
    get:
      parameters:
      - description: problem identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 问题详情
      tags:
      - 公共方法
  /problemList:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: category_identity
        in: query
        name: category_identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 获取多个标签
      tags:
      - 标签
  /rankList:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 用户排行榜
      tags:
      - 公共方法
  /register:
    post:
      parameters:
      - description: mail
        in: formData
        name: mail
        required: true
        type: string
      - description: code
        in: formData
        name: code
        required: true
        type: string
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 用户注册
      tags:
      - 公共方法
  /sendCode:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 发送验证码
      tags:
      - 公共方法
  /submitList:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: problemIdentity
        in: query
        name: problemIdentity
        type: string
      - description: userIdentity
        in: query
        name: userIdentity
        type: string
      - description: status
        in: query
        name: status
        type: integer
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 提交列表
      tags:
      - 公共方法
  /userDetail:
    get:
      parameters:
      - description: user identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":"","data":""}'
          schema:
            type: string
      summary: 用户详情
      tags:
      - 公共方法
swagger: "2.0"